function_codes/searchi_func/requirements.txt
azure-functions==1.19.0
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
urllib3==2.2.1
